[test]增加打开新窗口的测试用例 [build]升级selenium版本到4.1.2,以解决每次用例执行完调用service.close()或者driver.quit()后执行doubleDoubleClick用例报classNotFound(原因:doubleClick执行的时候会报一个UnsupportedCommandException,但driver或者service关闭后再重新初始化会导致这个class找不到) [refactor]sendKeysHandler重构 [feat] 1. jsUtil等待页面加载时候支持传入等待时间 2. waitHandler功能改成等待页面加载

+++ selenium-core/src/main/java/com/lml/selenium/dto/NoEleHandlerDto.java
@@ -16,9 +16,9 @@ import lombok.experimental.Accessors;
 public class NoEleHandlerDto extends BaseSeleniumDto {
 
     /**
-     * 等待的时间
+     * 等待的时间(单位:毫秒)
      */
-    private Integer waitTime;
+    private Long waitTime;
 
 
     /**
